requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.3
boto3==1.40.29
botocore==1.40.29
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
decorator==5.2.1
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
django-extensions==4.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
ipython==8.37.0
jedi==0.19.2
jiter==0.12.0
jmespath==1.0.1
kombu==5.5.4
matplotlib-inline==0.2.1
msgpack==1.1.2
openai==2.8.1
packaging==25.0
parso==0.8.5
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
redis==7.0.1
requests==2.32.5
s3transfer==0.14.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.3.0
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1
